+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - All compiled templates from one run will be stored in one directory within tmpDir
+- Feature phpstanCommand to analyse compiled templates via separated command  (Turn this feature with parameter `latte.features.phpstanCommand: "vendor/bin/phpstan {dir}"`)
 
 ## [0.15.0] - 2023-07-31
 ### Added